refactor:seperate commands to their own file
new folder within models for commands,
with subfolders for fs/config commands
